nested all of statistics_utils functions
