+++ woodwork/statistics_utils/_get_mode.py
@@ -1,0 +1,6 @@
+def _get_mode(series):
+    """Get the mode value for a series"""
+    mode_values = series.mode()
+    if len(mode_values) > 0:
+        return mode_values[0]
+    return None
